pyproject.toml
[tool.poetry]
name = "fft-signal-analyzer"
version = "1.0.0"
description = "FFT-based signal analyzer for frequency component extraction"
authors = ["Your Name <email@example.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yourusername/fft-signal-analyzer"
keywords = ["fft", "signal-processing", "dsp", "frequency-analysis", "filtering"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26"
scipy = "^1.12"
plotly = "^5.18"
dash = "^2.14"
dash-bootstrap-components = "^1.5"
pandas = "^2.1"
pyyaml = "^6.0"
click = "^8.1"
librosa = "^0.10"
fpdf2 = "^2.7"
streamlit = "^1.28"
streamlit-option-menu = "^0.3"
scikit-learn = "^1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-cov = "^4.1"
pytest-mock = "^3.12"
black = "^24.0"
ruff = "^0.2"
mypy = "^1.8"
types-PyYAML = "^6.0"
jupyter = "^1.0"
notebook = "^7.0"
pre-commit = "^3.6"
ipython = "^8.17"

[tool.poetry.scripts]
fft-analyzer = "fft_analyzer.cli.main:cli"
fft-dashboard = "fft_analyzer.ui.app:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.ruff.isort]
known-first-party = ["fft_analyzer"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = [
    "scipy.*",
    "plotly.*",
    "dash.*",
    "pandas.*",
    "yaml.*",
    "librosa.*",
    "fpdf.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["src/fft_analyzer"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
